Fix csup build issues on linux.

+++ usr.bin/csup/proto.c
@@ -754,7 +754,7 @@ proto_printf(struct stream *wr, const char *format, ...)
 			break;
 		case 'O':
 			off = va_arg(ap, off_t);
-			rv = stream_printf(wr, "%" PRId64, off);
+			rv = stream_printf(wr, "%" PRId64, (long long)off);
 			break;
 		case 'S':
 			s = va_arg(ap, char *);
